restructured the file directory of the notebook database
