--- cmd/app/models/user.go
@@ -1,13 +1,0 @@
-package models
-
-import (
-	"gorm.io/gorm"
-)
-
-type User struct {
-	gorm.Model
-	Email		string
-	Username	string
-	Password	string
-}
-
